📦 various

+++ src/StashBot/Handlers/CommandHandlers/ToolsCommandHandler.cs
@@ -31,6 +31,10 @@ namespace StashBot.Handlers.CommandHandlers
                     },
                     new []
                     {
+                        InlineKeyboardButton.WithCallbackData($"🔀 {Localization.GetPhrase(Localization.Phrase.RandomizeQueue, arguments.TelegramUser)}", $"tools_randomizequeue"),
+                    },
+                    new []
+                    {
                         InlineKeyboardButton.WithCallbackData($"🚽 {Localization.GetPhrase(Localization.Phrase.FlushRemovedPosts, arguments.TelegramUser)}", $"tools_flush"),
                     },
                     new []
@@ -52,6 +56,7 @@ namespace StashBot.Handlers.CommandHandlers
             else
             {
                 throw new CommandHandlerException(Localization.GetPhrase(Localization.Phrase.NoPermissionTools, arguments.TelegramUser));
+                //await HandleNoPermission(arguments);
             }
         }
 
@@ -71,73 +76,60 @@ namespace StashBot.Handlers.CommandHandlers
             }
             else
             {
-                await Program.BotClient.DeleteMessageAsync(
-                    chatId: arguments.TelegramCallbackQueryEvent.CallbackQuery.Message.Chat.Id,
-                    messageId: arguments.TelegramCallbackQueryEvent.CallbackQuery.Message.MessageId
-                );
-
-                throw new CommandHandlerException(Localization.GetPhrase(Localization.Phrase.NoPermissionTools, arguments.TelegramUser));
+                await HandleNoPermission(arguments);
             }
         }
 
-        public static void InvokePurgeUsers(CommandHandlerArguments arguments)
+        public async static void InvokePurgeUsers(CommandHandlerArguments arguments)
         {
             if (AuthorData.DoesAuthorExist(arguments.TelegramUser))
             {
                 if (AuthorData.CanAuthorManageAuthors(arguments.TelegramUser.Id))
                 {
-                    try
+                    List<Author> allAuthors = AuthorData.GetAuthors();
+                    List<Author> authorsToDelete = new List<Author>();
+
+                    foreach (var author in allAuthors)
                     {
-                        List<Author> allAuthors = AuthorData.GetAuthors();
-                        List<Author> authorsToDelete = new List<Author>();
+                        int queueItemAmount = author.QueueItems.Count;
 
-                        foreach (var author in allAuthors)
+                        if (
+                            author.CanDeleteOthers == false &&
+                            author.CanFlushQueue == false &&
+                            author.CanManageAuthors == false &&
+                            author.CanQueue == false &&
+                            queueItemAmount == 0
+                        )
                         {
-                            int queueItemAmount = author.QueueItems.Count;
-
-                            if (
-                                author.CanDeleteOthers == false &&
-                                author.CanFlushQueue == false &&
-                                author.CanManageAuthors == false &&
-                                author.CanQueue == false &&
-                                queueItemAmount == 0
-                            )
-                            {
-                                authorsToDelete.Add(author);
-                            }
-                        }
-
-                        if (authorsToDelete.Count == 0)
-                        {
-                            MessageUtilities.AlertWarningMessage(Localization.GetPhrase(Localization.Phrase.NoDanglingUsers, arguments.TelegramUser), arguments.TelegramCallbackQueryEvent);
-                        }
-                        else
-                        {
-                            AuthorData.DeleteAuthorRange(authorsToDelete);
-                            MessageUtilities.AlertSuccessMessage(Localization.GetPhrase(
-                                Localization.Phrase.FlushedXDanglingUsers,
-                                arguments.TelegramUser,
-                                new string[] {
-                                        authorsToDelete.Count.ToString()
-                                }
-                            ),
-                            arguments.TelegramCallbackQueryEvent);
+                            authorsToDelete.Add(author);
                         }
                     }
-                    catch (Exception e)
+
+                    if (authorsToDelete.Count == 0)
                     {
-                        MessageUtilities.SendErrorMessage(e, arguments.TelegramCallbackQueryEvent);
+                        MessageUtilities.AlertWarningMessage(Localization.GetPhrase(Localization.Phrase.NoDanglingUsers, arguments.TelegramUser), arguments.TelegramCallbackQueryEvent);
+                    }
+                    else
+                    {
+                        AuthorData.DeleteAuthorRange(authorsToDelete);
+                        MessageUtilities.AlertSuccessMessage(Localization.GetPhrase(
+                            Localization.Phrase.FlushedXDanglingUsers,
+                            arguments.TelegramUser,
+                            new string[] {
+                                        authorsToDelete.Count.ToString()
+                            }
+                        ),
+                        arguments.TelegramCallbackQueryEvent);
                     }
                 }
                 else
                 {
                     MessageUtilities.AlertWarningMessage(Localization.GetPhrase(Localization.Phrase.NoPermissionFlushDanglingUsers, arguments.TelegramUser), arguments.TelegramCallbackQueryEvent);
-                    //throw new CommandHandlerAlertException("You do not have permission to purge dangling users");
                 }
             }
             else
             {
-                throw new CommandHandlerException(Localization.GetPhrase(Localization.Phrase.NoPermissionTools, arguments.TelegramUser));
+                await HandleNoPermission(arguments);
             }
         }
 
@@ -157,13 +149,46 @@ namespace StashBot.Handlers.CommandHandlers
             }
             else
             {
+                await HandleNoPermission(arguments);
+            }
+        }
+
+        public async static Task InvokeRandomizeQueue(CommandHandlerArguments arguments)
+        {
+            if (AuthorData.DoesAuthorExist(arguments.TelegramUser))
+            {
+                if (AuthorData.CanAuthorRandomizeQueue(arguments.TelegramUser.Id))
+                {
+                    Constants.IsPostingDisabled = true;
+                    QueueData.RandomizeQueuedQueueItems();
+                    Constants.IsPostingDisabled = false;
+                    MessageUtilities.AlertSuccessMessage(Localization.GetPhrase(Localization.Phrase.RandomizedQueue, arguments.TelegramUser), arguments.TelegramCallbackQueryEvent);
+                }
+                else
+                {
+                    MessageUtilities.AlertWarningMessage(Localization.GetPhrase(Localization.Phrase.NoPermissionRandomizeQueue, arguments.TelegramUser), arguments.TelegramCallbackQueryEvent);
+                }
+            }
+            else
+            {
+                await HandleNoPermission(arguments);
+            }
+        }
+
+        private async static Task HandleNoPermission(
+            CommandHandlerArguments arguments,
+            Localization.Phrase phrase = Localization.Phrase.NoPermissionTools
+        )
+        {
+            if (arguments.TelegramCallbackQueryEvent != null)
+            {
                 await Program.BotClient.DeleteMessageAsync(
                     chatId: arguments.TelegramCallbackQueryEvent.CallbackQuery.Message.Chat.Id,
                     messageId: arguments.TelegramCallbackQueryEvent.CallbackQuery.Message.MessageId
                 );
-
-                throw new CommandHandlerException(Localization.GetPhrase(Localization.Phrase.NoPermissionTools, arguments.TelegramUser));
             }
+
+            throw new CommandHandlerException(Localization.GetPhrase(phrase, arguments.TelegramUser));
         }
     }
 }
